requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==22.0.0

# ── OpenAI ───────────────────────────────────────────────────────────────
openai==1.58.1

# ── Database ─────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# ── Redis ────────────────────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── Vector Store ─────────────────────────────────────────────────────────
faiss-cpu==1.9.0.post1
numpy==1.26.4

# ── Background Tasks ────────────────────────────────────────────────────
celery[redis]==5.4.0
flower==2.0.1

# ── Config ───────────────────────────────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.1

# ── Logging ──────────────────────────────────────────────────────────────
structlog==24.4.0

# ── Metrics ──────────────────────────────────────────────────────────────
prometheus-client==0.21.0

# ── Rate Limiting ────────────────────────────────────────────────────
slowapi==0.1.9

# ── Testing ──────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-env==1.1.5
httpx==0.28.1
anyio==4.8.0
